Update to groovy 2.3.4

+++ src/main/groovy/graxxia/Stats.groovy
@@ -234,7 +234,7 @@ class Stats extends DescriptiveStatistics {
     @CompileStatic
     static SummaryStatistics summary(double [] values) {
         int i=0;
-        summary { values[i++] }
+        summary { values[(int)i++] }
     }
     
     /**
